+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 3.0.0
 =====
 
+*   (bc) Require Symfony 5 and update config.
 *   (bug) Properly support Symfony 5.
 *   (improvement) Bump required packages.
 *   (internal) Bump sentry packages.